requirements.txt
docutils==0.14
Pillow==6.0.0

dj-database-url==0.5.0
Django==2.2.2
django-heroku==0.3.1
django-crispy-forms==1.7.2
django-debug-toolbar==2.0

gunicorn==19.9.0
psycopg2-binary==2.8.2
whitenoise==4.1.2

djangorestframework==3.9.4
django-cors-headers==3.0.2

django-import-export==1.2.0


